refactor(package): change package structure
