+++ src/main/java/blackjack/domain/participant/Participants.java
@@ -1,5 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Deck;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
