+++ src/main/java/blackjack/domain/result/WinningResult.java
@@ -1,4 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.result;
+
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Player;
 
 public enum WinningResult {
